Remove events and replace with models
