+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod application;
 pub mod configuration;
-pub mod events;
 pub mod graphql;
+pub mod models;
 pub mod routes;
 pub mod telemetry;
